requirements.txt
requests>=2.31.0
pandas>=2.0.0
python-dotenv>=1.0.0
pyodbc>=4.0.39
openpyxl>=3.1.2
xlwings>=0.30.10
selenium>=4.15.0
webdriver-manager>=4.0.0 